+++ History.md
@@ -1,10 +1,22 @@
 Major Revisions
 ===============
 
+1.0.0 / 2016-11-06
+------------------
+
+- Breaking changes to wit interface. 
+  - The wit message service is now accessible at `require('wit/message')` so change `require('wit')` to `require('wit/message')` after upgrade.
+- Added `wit.converse` or `require('wit/converse')`
+- Removed deprecated `wit.message`
+- Removed deprecated `conversation.message`
+
+
 0.2.1 / 2016-11-05
 ------------------
 
--	Modernise conversation and wit services. Further simplify their interface
+- Modernise conversation and wit services. Further simplify their interface
+  - deprecate usage of `wit.message()`, instead use `wit()` directly.
+  - deprecate usage of `conversation.message()`, instead use `conversation()` directly.
 
 
 0.2.0 / 2016-11-05